WIP refactoring for single responsibility with testing including multiformats which impacted all imports to require .js extensions.
