+++ package.json
@@ -2,6 +2,7 @@
   "name": "instagramtobluesky",
   "version": "0.3.0",
   "description": "Import Instagram archive to a Bluesky account",
+  "type": "module",
   "main": "app.js",
   "engines": {
     "node": ">=20.12.0"
@@ -17,13 +18,16 @@
   "dependencies": {
     "@atproto/api": "^0.13.31",
     "@ffprobe-installer/ffprobe": "^2.1.2",
+    "byte-size": "^9.0.1",
     "dotenv": "^16.4.7",
     "fluent-ffmpeg": "^2.1.3",
     "luxon": "^3.5.0",
     "multiformats": "^13.3.2",
+    "multihashes": "^4.0.3",
     "pino": "^9.6.0",
     "pino-pretty": "^13.0.0",
-    "process": "^0.11.10"
+    "process": "^0.11.10",
+    "sharp": "^0.33.5"
   },
   "devDependencies": {
     "@types/jest": "^29.5.14",
