+++ src/bluesky/index.ts
@@ -1,2 +1,2 @@
-export * from './types';
-export * from './bluesky';
+export * from './types/index.js';
+export * from './bluesky.js';
