+++ src/bluesky/types/EmbeddedMedia.ts
@@ -1,4 +1,4 @@
-import { VideoEmbed } from "./VideoEmbed";
-import { ImagesEmbed } from "./ImagesEmbed";
+import { VideoEmbed } from "./VideoEmbed.js";
+import { ImagesEmbed } from "./ImagesEmbed.js";
 
 export type EmbeddedMedia = VideoEmbed | ImagesEmbed;
