+++ src/bluesky/types/index.ts
@@ -1,5 +1,5 @@
-export * from "./ImageEmbed";
-export * from "./VideoEmbed";
-export * from "./ImagesEmbed";
-export * from "./PostRecord";
-export * from "./EmbeddedMedia"; 
+export * from "./ImageEmbed.js";
+export * from "./VideoEmbed.js";
+export * from "./ImagesEmbed.js";
+export * from "./PostRecord.js";
+export * from "./EmbeddedMedia.js";
