+++ src/image/index.ts
@@ -1,1 +1,1 @@
-export * from './image';
+export * from './image.js';
