+++ src/main.ts
@@ -1,4 +1,4 @@
-import { main } from "./instagram-to-bluesky";
+import { main } from "./instagram-to-bluesky.js";
 
 (async () => {
     await main();
