+++ src/media/index.ts
@@ -1,1 +1,3 @@
-export * from './media';
+export * from './MediaProcessResult.js';
+export * from './ProcessedPost.js';
+export * from './media.js';
