+++ src/video/index.ts
@@ -1,1 +1,1 @@
-export * from './video';
+export * from './video.js';
